third_party/blink/web_tests/fast/forms/datalist/input-list-expected.txt

Test for the list attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.list is null
PASS input.list is null
PASS input.list is null
PASS datalist.className is "former"
PASS document.getElementById("text").list is document.getElementById("dl1")
PASS document.getElementById("search").list is document.getElementById("dl1")
PASS document.getElementById("url").list is document.getElementById("dl1")
PASS document.getElementById("telephone").list is document.getElementById("dl1")
PASS document.getElementById("email").list is document.getElementById("dl1")
PASS document.getElementById("datetime").list is document.getElementById("dl1")
PASS document.getElementById("date").list is document.getElementById("dl1")
PASS document.getElementById("month").list is document.getElementById("dl1")
PASS document.getElementById("week").list is document.getElementById("dl1")
PASS document.getElementById("time").list is document.getElementById("dl1")
PASS document.getElementById("datetime-local").list is document.getElementById("dl1")
PASS document.getElementById("number").list is document.getElementById("dl1")
PASS document.getElementById("range").list is document.getElementById("dl1")
PASS document.getElementById("color").list is document.getElementById("dl1")
PASS document.getElementById("hidden").list is null
PASS document.getElementById("password").list is null
PASS document.getElementById("checkbox").list is null
PASS document.getElementById("radio").list is null
PASS document.getElementById("file").list is null
PASS document.getElementById("submit").list is null
PASS document.getElementById("image").list is null
PASS document.getElementById("reset").list is null
PASS document.getElementById("button").list is null
PASS successfullyParsed is true

TEST COMPLETE
